refactoring using inner record for LineItem
