--- samples/spring-eventsourced-shopping-cart/src/main/java/com/example/shoppingcart/domain/LineItem.java
@@ -1,7 +1,0 @@
-package com.example.shoppingcart.domain;
-
-public record LineItem(String productId, String name, int quantity) {
-  public LineItem withQuantity(int quantity) {
-    return new LineItem(productId, name, quantity);
-  }
-}
